Feat: integration with Add Recipe

+++ Final_RecipeLab_Team16/Schema/Recipe.swift
@@ -5,7 +5,7 @@ struct Recipe: Identifiable, Codable {
     @DocumentID var id: String?
     let title: String
     let imageUrl: String
-    let cookingTime: Int8
+    let cookingTime: Int
     let ingredients: [String]
     let steps: [String]
     let userName: String
